check for authorization

+++ app/Http/Controllers/ArticleController.php
@@ -5,9 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Article;
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware("auth")->except(['index', 'detail']);
+    }
     public function index()
     {
         $data = Article::latest()->paginate(5);
@@ -29,9 +34,12 @@ class ArticleController extends Controller
     {
         $article = Article::find($id);
 
-        $article->delete();
+        if (Gate::allows('article-delete', $article)) {
+            $article->delete();
+            return redirect("/articles")->with('info', 'An article deleted!');
+        }
 
-        return redirect("/articles")->with('info', 'An article deleted!');
+        return redirect("/articles")->with('info', 'Unauthorized to Deleted!');
     }
 
     public function add()
@@ -58,6 +66,38 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
+        $article->save();
+
+        return redirect("/articles")->with("info", "An article added!");
+    }
+    public function edit($id)
+    {
+        $article = Article::find($id);
+        $categories = Category::all();
+
+        return view("articles.edit", [
+            "article" => $article,
+            "categories" => $categories
+        ]);
+    }
+    public function update($id)
+    {
+        $validator = validator(request()->all(), [
+            'title' => 'required',
+            'body' => 'required',
+            'category_id' => 'required',
+        ]);
+
+        if($validator->fails()) {
+            return back()->withErrors($validator);
+        }
+
+        $article = Article::find($id);
+        $article->title = request()->title;
+        $article->body = request()->body;
+        $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
         $article->save();
 
         return redirect("/articles")->with("info", "An article added!");
