+++ app/Http/Controllers/CommentController.php
@@ -4,9 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware("auth");
+    }
     public function create()
     {
         $validator = validator(request()->all(), [
@@ -19,6 +24,7 @@ class CommentController extends Controller
         $comment = new Comment();
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back()->with('info', 'A comment Added!');
@@ -26,8 +32,11 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
+        if (Gate::allows("comment-delete", $comment)) {
+            $comment->delete();
+            return back()->with('info', 'A comment Deleted!');
+        }
 
-        return back()->with('info', 'A comment Deleted!');
+        return back()->with('info', 'Unathorized to Deleted!');
     }
 }
